refactor(talk): move Apple relay into OpenClawKit
