+++ apps/ios/Sources/Voice/TalkModeManager.swift
@@ -2321,6 +2321,10 @@ final class TalkModeManager: NSObject {
             return .ignored
         }
         guard self.isCurrentStartAttempt(attemptID) else { return .ignored }
+        guard let gatewayRoute = await gateway.currentRoute() else {
+            return .unavailable(realtimeIssue(message: "Gateway not connected", phase: "start"))
+        }
+        guard self.isCurrentStartAttempt(attemptID) else { return .ignored }
         if self.realtimeRelaySession != nil {
             GatewayDiagnostics.log("talk realtime ignored: already active")
             return .started
@@ -2342,19 +2346,27 @@ final class TalkModeManager: NSObject {
         GatewayDiagnostics.log("talk.timeline realtime relay start attempt sessionKey=\(sessionKey)")
         let startedAt = Self.nowSeconds()
         let relaySession = RealtimeTalkRelaySession(
-            gateway: gateway,
+            transport: .ios(gateway: gateway, route: gatewayRoute),
             options: RealtimeTalkRelaySession.Options(
                 sessionKey: sessionKey,
                 provider: self.realtimeProvider,
                 model: self.realtimeModelId,
                 voice: self.realtimeVoiceId),
+            audioCapture: IOSRealtimeTalkAudioCapture(),
             pcmPlayer: self.pcmPlayer,
             onStatus: { [weak self] status in
                 guard let self, self.realtimeRelayGeneration == relayGeneration else { return }
                 self.handleRealtimeRelayStatus(status)
             },
-            onIssue: { [weak self] issue in
+            onIssue: { [weak self] relayIssue in
                 guard let self, self.realtimeRelayGeneration == relayGeneration else { return }
+                let issue = TalkRuntimeIssue(
+                    code: .realtimeUnavailable,
+                    message: relayIssue.message,
+                    provider: relayIssue.provider,
+                    model: relayIssue.model,
+                    transport: relayIssue.transport,
+                    phase: relayIssue.phase)
                 self.realtimeRelayStartIssue = issue
                 self.pendingRealtimeIssue = issue
                 self.gatewayTalkLastIssueText = issue.diagnosticSummary
