+++ apps/shared/OpenClawKit/Tests/OpenClawKitTests/RealtimeTalkRelaySessionTests.swift
@@ -1,0 +1,428 @@
+import Foundation
+import OpenClawKit
+import OpenClawProtocol
+import Testing
+@testable import OpenClawKit
+
+@MainActor
+private final class UnusedPCMStreamingAudioPlayer: PCMStreamingAudioPlaying {
+    func play(stream: AsyncThrowingStream<Data, Error>, sampleRate: Double) async -> StreamingPlaybackResult {
+        fatalError("Playback is not used by this test")
+    }
+
+    func stop() -> Double? {
+        nil
+    }
+}
+
+@MainActor
+private final class TestRealtimeTalkAudioCapture: RealtimeTalkAudioCapturing {
+    var suppressesInputDuringOutput = false
+    private(set) var isStarted = false
+    private(set) var startCount = 0
+    private(set) var stopCount = 0
+
+    func start(
+        targetSampleRate: Double,
+        onAudio: @escaping @Sendable (RealtimeTalkAudioFrame) -> Void) throws
+    {
+        self.isStarted = true
+        self.startCount += 1
+    }
+
+    func stop() {
+        self.isStarted = false
+        self.stopCount += 1
+    }
+}
+
+private actor RealtimeRelayStartupBarrier {
+    private var entered = false
+    private var enteredWaiter: CheckedContinuation<Void, Never>?
+    private var releaseWaiter: CheckedContinuation<Void, Never>?
+
+    func suspend() async {
+        self.entered = true
+        self.enteredWaiter?.resume()
+        self.enteredWaiter = nil
+        await withCheckedContinuation { self.releaseWaiter = $0 }
+    }
+
+    func waitUntilEntered() async {
+        if self.entered {
+            return
+        }
+        await withCheckedContinuation { self.enteredWaiter = $0 }
+    }
+
+    func release() {
+        self.releaseWaiter?.resume()
+        self.releaseWaiter = nil
+    }
+}
+
+private struct RealtimeRelayStartupRequest: Sendable {
+    let method: String
+    let params: [String: AnyCodable]?
+}
+
+private actor RealtimeRelayStartupRequestLog {
+    private var requests: [RealtimeRelayStartupRequest] = []
+
+    func record(method: String, params: [String: AnyCodable]?) {
+        self.requests.append(RealtimeRelayStartupRequest(method: method, params: params))
+    }
+
+    func snapshot() -> [RealtimeRelayStartupRequest] {
+        self.requests
+    }
+}
+
+private func unusedRealtimeRelayTransport() -> RealtimeTalkRelayTransport {
+    RealtimeTalkRelayTransport(
+        subscribeServerEvents: { _ in AsyncStream { $0.finish() } },
+        request: { _, _, _ in throw CancellationError() })
+}
+
+@MainActor
+struct RealtimeTalkRelaySessionTests {
+    @Test func `transcript callback carries typed partial and final values`() async {
+        var transcripts: [RealtimeTalkTranscript] = []
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: nil, model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in },
+            onTranscript: { transcripts.append($0) })
+        session._test_setRelaySessionId("relay-1")
+
+        for isFinal in [false, true] {
+            await session._test_handleGatewayEvent(EventFrame(
+                type: "event",
+                event: "talk.event",
+                payload: AnyCodable([
+                    "relaySessionId": "relay-1",
+                    "type": "transcript",
+                    "role": "user",
+                    "text": isFinal ? "hello" : "hel",
+                    "final": isFinal,
+                ]),
+                seq: nil,
+                stateversion: nil))
+        }
+
+        #expect(transcripts == [
+            RealtimeTalkTranscript(role: "user", text: "hel", isFinal: false),
+            RealtimeTalkTranscript(role: "user", text: "hello", isFinal: true),
+        ])
+    }
+
+    @Test func `input pause and resume are idempotent and keep relay alive`() throws {
+        let audioCapture = TestRealtimeTalkAudioCapture()
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: nil, model: nil, voice: nil),
+            audioCapture: audioCapture,
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in })
+        session._test_setRelaySessionId("relay-1")
+
+        try session.setInputPaused(true)
+        try session.setInputPaused(true)
+        try session.setInputPaused(false)
+        try session.setInputPaused(false)
+
+        #expect(audioCapture.stopCount == 2)
+        #expect(audioCapture.startCount == 1)
+        #expect(audioCapture.isStarted)
+    }
+
+    @Test func `output playback finish clears barge in start time`() {
+        var speakingStates: [Bool] = []
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: nil, model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { speakingStates.append($0) })
+
+        session._test_markOutputAudioStarted(nowMs: 100)
+        #expect(session._test_isOutputPlaying())
+        #expect(session._test_outputStartedAtMs() == 100)
+
+        session._test_markOutputPlaybackFinished()
+        #expect(!session._test_isOutputPlaying())
+        #expect(session._test_outputStartedAtMs() == nil)
+        #expect(speakingStates == [false])
+
+        session._test_markOutputAudioStarted(nowMs: 500)
+        #expect(session._test_outputStartedAtMs() == 500)
+    }
+
+    @Test func `playback mark is acknowledged after output finishes`() async throws {
+        let requests = RealtimeRelayStartupRequestLog()
+        let transport = RealtimeTalkRelayTransport(
+            subscribeServerEvents: { _ in AsyncStream { $0.finish() } },
+            request: { method, params, _ in
+                await requests.record(method: method, params: params)
+                return Data("{\"ok\":true}".utf8)
+            })
+        let session = RealtimeTalkRelaySession(
+            transport: transport,
+            options: .init(sessionKey: "main", provider: "xai", model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in })
+        session._test_setRelaySessionId("relay-1")
+        session._test_markOutputAudioStarted(nowMs: 100)
+
+        await session._test_handleGatewayEvent(EventFrame(
+            type: "event",
+            event: "talk.event",
+            payload: AnyCodable([
+                "relaySessionId": "relay-1",
+                "type": "mark",
+                "markName": "audio-1",
+            ]),
+            seq: nil,
+            stateversion: nil))
+        await Task.yield()
+        #expect(await requests.snapshot().isEmpty)
+
+        session._test_markOutputPlaybackFinished()
+        for _ in 0..<10 {
+            if !(await requests.snapshot()).isEmpty { break }
+            await Task.yield()
+        }
+
+        let recorded = await requests.snapshot()
+        #expect(recorded.count == 1)
+        let request = try #require(recorded.first)
+        #expect(request.method == "talk.session.acknowledgeMark")
+        #expect(request.params?["sessionId"]?.stringValue == "relay-1")
+        #expect(request.params?["markName"]?.stringValue == "audio-1")
+    }
+
+    @Test func `close after classified error does not replace issue`() async {
+        var issues: [RealtimeTalkRelayIssue] = []
+        var statuses: [String] = []
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: "openai", model: "gpt-realtime-2", voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { statuses.append($0) },
+            onIssue: { issues.append($0) },
+            onSpeakingChanged: { _ in })
+        session._test_setRelaySessionId("relay-1")
+
+        await session._test_handleGatewayEvent(EventFrame(
+            type: "event",
+            event: "talk.event",
+            payload: AnyCodable([
+                "relaySessionId": "relay-1",
+                "type": "error",
+                "message": "OpenAI API key rejected with 401",
+                "code": "realtime_unavailable",
+                "provider": "openai",
+                "model": "gpt-realtime-2",
+                "transport": "gateway-relay",
+                "phase": "connect",
+            ]),
+            seq: nil,
+            stateversion: nil))
+        await session._test_handleGatewayEvent(EventFrame(
+            type: "event",
+            event: "talk.event",
+            payload: AnyCodable([
+                "relaySessionId": "relay-1",
+                "type": "close",
+                "reason": "error",
+            ]),
+            seq: nil,
+            stateversion: nil))
+
+        #expect(issues.map(\.code) == ["realtime_unavailable"])
+        #expect(statuses == ["OpenAI API key rejected with 401"])
+    }
+
+    @Test func `closed relay does not wait for startup ready`() async {
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: "openai", model: "gpt-realtime-2", voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in })
+
+        session.stop()
+
+        #expect(await session._test_waitForStartupCancelled(timeoutSeconds: 1))
+    }
+
+    @Test func `startup ready wait covers gateway connect budget`() {
+        let session = RealtimeTalkRelaySession(
+            transport: unusedRealtimeRelayTransport(),
+            options: .init(sessionKey: "main", provider: "openai", model: "gpt-realtime-2", voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in })
+
+        #expect(session._test_startupReadyTimeoutSeconds() >= 12)
+    }
+
+    @Test func `stop during event subscription prevents relay creation`() async throws {
+        let barrier = RealtimeRelayStartupBarrier()
+        let requests = RealtimeRelayStartupRequestLog()
+        var statuses: [String] = []
+        var speakingStates: [Bool] = []
+        let transport = RealtimeTalkRelayTransport(
+            subscribeServerEvents: { _ in
+                await barrier.suspend()
+                return AsyncStream { $0.finish() }
+            },
+            request: { method, params, _ in
+                await requests.record(method: method, params: params)
+                throw URLError(.badServerResponse)
+            })
+        let session = RealtimeTalkRelaySession(
+            transport: transport,
+            options: .init(sessionKey: "main", provider: "openai", model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { statuses.append($0) },
+            onSpeakingChanged: { speakingStates.append($0) })
+        let start = Task { @MainActor in try await session.start() }
+        await barrier.waitUntilEntered()
+
+        session.stop()
+        await barrier.release()
+        try await start.value
+
+        #expect(await requests.snapshot().isEmpty)
+        #expect(statuses == ["Connecting realtime…"])
+        #expect(!speakingStates.contains(true))
+    }
+
+    @Test func `stop during relay creation closes late session once`() async throws {
+        let barrier = RealtimeRelayStartupBarrier()
+        let requests = RealtimeRelayStartupRequestLog()
+        var statuses: [String] = []
+        var speakingStates: [Bool] = []
+        let result = TalkSessionCreateResult(
+            sessionid: "talk-session",
+            mode: AnyCodable("realtime"),
+            transport: AnyCodable("gateway-relay"),
+            brain: AnyCodable("agent-consult"),
+            relaysessionid: "relay-1")
+        let resultData = try JSONEncoder().encode(result)
+        let transport = RealtimeTalkRelayTransport(
+            subscribeServerEvents: { _ in AsyncStream { $0.finish() } },
+            request: { method, params, _ in
+                await requests.record(method: method, params: params)
+                if method == "talk.session.create" {
+                    await barrier.suspend()
+                    return resultData
+                }
+                return Data("{}".utf8)
+            })
+        let session = RealtimeTalkRelaySession(
+            transport: transport,
+            options: .init(sessionKey: "main", provider: "openai", model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { statuses.append($0) },
+            onSpeakingChanged: { speakingStates.append($0) })
+        let start = Task { @MainActor in try await session.start() }
+        await barrier.waitUntilEntered()
+
+        session.stop()
+        await barrier.release()
+        try await start.value
+
+        let recorded = await requests.snapshot()
+        #expect(recorded.map(\.method) == ["talk.session.create", "talk.session.close"])
+        #expect(recorded.last?.params?["sessionId"]?.stringValue == "relay-1")
+        #expect(!statuses.contains("Waiting for realtime…"))
+        #expect(!speakingStates.contains(true))
+    }
+
+    @Test func `stop during buffered tool call prevents late relay side effects`() async {
+        let barrier = RealtimeRelayStartupBarrier()
+        let requests = RealtimeRelayStartupRequestLog()
+        var statuses: [String] = []
+        let transport = RealtimeTalkRelayTransport(
+            subscribeServerEvents: { _ in AsyncStream { $0.finish() } },
+            request: { method, params, _ in
+                await requests.record(method: method, params: params)
+                if method == "talk.client.toolCall" {
+                    await barrier.suspend()
+                    return Data("{\"runId\":\"run-1\"}".utf8)
+                }
+                return Data("{}".utf8)
+            })
+        let session = RealtimeTalkRelaySession(
+            transport: transport,
+            options: .init(sessionKey: "main", provider: "openai", model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { statuses.append($0) },
+            onSpeakingChanged: { _ in })
+        session._test_setRelaySessionId("relay-1")
+        let handling = Task { @MainActor in
+            await session._test_handleGatewayEvent(EventFrame(
+                type: "event",
+                event: "talk.event",
+                payload: AnyCodable([
+                    "relaySessionId": "relay-1",
+                    "type": "toolCall",
+                    "callId": "call-1",
+                    "name": "lookup",
+                    "args": [:],
+                ]),
+                seq: nil,
+                stateversion: nil))
+        }
+        await barrier.waitUntilEntered()
+
+        session.stop()
+        await barrier.release()
+        await handling.value
+        await session._test_waitForToolCalls()
+
+        let methods = await requests.snapshot().map(\.method)
+        #expect(methods.first == "talk.client.toolCall")
+        #expect(!methods.contains("talk.session.submitToolResult"))
+        #expect(statuses == ["Thinking…"])
+    }
+
+    @Test func `stop cancels buffered microphone audio before dispatch`() async throws {
+        let requests = RealtimeRelayStartupRequestLog()
+        let transport = RealtimeTalkRelayTransport(
+            subscribeServerEvents: { _ in AsyncStream { $0.finish() } },
+            request: { method, params, _ in
+                await requests.record(method: method, params: params)
+                return Data("{\"ok\":true}".utf8)
+            })
+        let session = RealtimeTalkRelaySession(
+            transport: transport,
+            options: .init(sessionKey: "main", provider: "openai", model: nil, voice: nil),
+            audioCapture: TestRealtimeTalkAudioCapture(),
+            pcmPlayer: UnusedPCMStreamingAudioPlayer(),
+            onStatus: { _ in },
+            onSpeakingChanged: { _ in })
+        session._test_prepareAudioSender(relaySessionId: "relay-1")
+        let send = try #require(session._test_enqueueMicrophoneFrame(Data([0x01, 0x02])))
+
+        session.stop()
+        await send.value
+
+        #expect(await requests.snapshot().isEmpty)
+    }
+}
